+++ Runtime/Scripts/KtxTexture.cs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections;
 using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.Experimental.Rendering;
@@ -20,43 +19,80 @@ using UnityEngine.Profiling;
 using Unity.Collections;
 
 namespace KtxUnity {
-    public class KtxTexture : TextureBase
-    {
-        public override async Task<TextureResult> LoadBytesRoutine(NativeSlice<byte> data, bool linear = false) {
+    public class KtxTexture : TextureBase {
 
-            var ktx = new KtxNativeInstance(data);
-            TextureResult result;
+        KtxNativeInstance m_Ktx;
 
-            if(ktx.valid) {
-                if(ktx.ktxClass==KtxClassId.ktxTexture2_c) {
-                    if(ktx.needsTranscoding) {
-                        result = await Transcode(ktx,linear);
-                        result.orientation = ktx.orientation;
+        public override ErrorCode Load(NativeSlice<byte> data) {
+            m_Ktx = new KtxNativeInstance();
+            return m_Ktx.Load(data);
+        }
+        
+        public override async Task<ErrorCode> Transcode(bool linear = false) {
+
+            ErrorCode result;
+
+            if(m_Ktx.valid) {
+                if(m_Ktx.ktxClass==KtxClassId.ktxTexture2_c) {
+                    if(m_Ktx.needsTranscoding) {
+                        result = await TranscodeInternal(m_Ktx,linear);
+                        // result.orientation = ktx.orientation;
                     } else {
-                        result = new TextureResult(ErrorCode.NotSuperCompressed);
+                        result = ErrorCode.NotSuperCompressed;
                     }
                 } else {
-                    result = new TextureResult(ErrorCode.UnsupportedVersion);
+                    result = ErrorCode.UnsupportedVersion;
                 }
             } else {
-                result = new TextureResult(ErrorCode.LoadingFailed);
+                result = ErrorCode.LoadingFailed;
             }
-            ktx.Unload();
             return result;
         }
 
-        async Task<TextureResult> Transcode(KtxNativeInstance ktx, bool linear) {
+        public override TextureResult CreateTexture() {
+            TextureResult result;
+            Profiler.BeginSample("LoadBytesRoutineGpuUpload");
+            try {
+                var texture = m_Ktx.LoadTextureData(m_Format);
+                result = new TextureResult {
+                    texture = texture
+                };
+            }
+            catch (UnityException) {
+                result = new TextureResult(ErrorCode.TranscodeFailed);
+            }
+            
+            Profiler.EndSample();
+            return result;
+        }
+        
+        public override void Dispose() {
+            m_Ktx.Unload();
+        }
+
+        async Task<ErrorCode> TranscodeInternal(KtxNativeInstance ktx, bool linear) {
             // TODO: Maybe do this somewhere more central
             TranscodeFormatHelper.Init();
 
-            TextureResult result = null;
+            var result = ErrorCode.Success;
             
             var formats = GetFormat(ktx,ktx,linear);
 
             if(formats.HasValue) {
-                var gf = formats.Value.format;
+                m_Format = formats.Value.format;
+
+                if(m_Format== GraphicsFormat.RGBA_DXT5_SRGB && !m_Ktx.hasAlpha) {
+                    // ktx library automatically decides to use the smaller
+                    // DXT1 instead of DXT5 if no alpha channel is present
+#if UNITY_2018_3_OR_NEWER
+                    m_Format = GraphicsFormat.RGBA_DXT1_SRGB;
+#else
+                    m_Format = GraphicsFormat.RGB_DXT1_SRGB;
+#endif
+                }
+
 #if KTX_VERBOSE
-                Debug.LogFormat("Transcode to GraphicsFormat {0} ({1})",gf,formats.Value.transcodeFormat);
+                Debug.LogFormat("Transcode to GraphicsFormat {0} ({1})",m_Format,formats.Value.transcodeFormat);
 #endif
                 Profiler.BeginSample("KtxTranscode");
 
@@ -74,33 +110,12 @@ namespace KtxUnity {
                 }
                 jobHandle.Complete();
 
-                if(job.result[0] == KtxErrorCode.KTX_SUCCESS) {
-                    Profiler.BeginSample("LoadBytesRoutineGPUupload");
-
-                    if(formats.Value.format== GraphicsFormat.RGBA_DXT5_SRGB && !ktx.hasAlpha) {
-                        // ktx library automatically decides to use the smaller DXT1 instead of DXT5 if no alpha
-#if UNITY_2018_3_OR_NEWER
-                        gf = GraphicsFormat.RGBA_DXT1_SRGB;
-#else
-                        gf = GraphicsFormat.RGB_DXT1_SRGB;
-#endif
-                    }
-                    try {
-                        var texture = ktx.LoadTextureData(gf);
-                        result = new TextureResult {
-                            texture = texture
-                        };
-                    }
-                    catch (UnityException) {
-                        result = new TextureResult(ErrorCode.TranscodeFailed);
-                    }
-                    Profiler.EndSample();
-                } else {
-                    result = new TextureResult(ErrorCode.TranscodeFailed);
+                if(job.result[0] != KtxErrorCode.KTX_SUCCESS) {
+                    result = ErrorCode.TranscodeFailed;
                 }
                 job.result.Dispose();
             } else {
-                result = new TextureResult(ErrorCode.UnsupportedFormat);
+                result = ErrorCode.UnsupportedFormat;
             }
             return result;
         }