+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+### Added
+- Low-level API that gives finer control over the loading process (alternative to `TextureBase.LoadBytesRoutine`)
+
 ## [1.2.3] - 2022-04-14
 ### Fixed
 - Fix undefined variable error when building project